break pubsub grid up
client and worker are no longer in the same file.  I moved everything to
be under a `pubsub/` folder
