+++ grid/__init__.py
@@ -1,4 +1,5 @@
 from grid.bygone import bygone
 from grid.client.grid import Grid as GridClient
 from grid.server.grid import Server as GridWorker
+from grid.pubsub import *
 from grid.lib import *
